+++ src/Digdir.Domain.Dialogporten.Application/Common/Behaviours/SilentUpdateBeaviour.cs
@@ -1,7 +1,7 @@
+using Digdir.Domain.Dialogporten.Application.Common.Context;
 using Digdir.Domain.Dialogporten.Application.Common.Extensions;
 using Digdir.Domain.Dialogporten.Application.Common.ReturnTypes;
 using Digdir.Domain.Dialogporten.Domain.Common;
-using Digdir.Domain.Dialogporten.Domain.Common.DomainEvents;
 using Digdir.Domain.Dialogporten.Domain.Common.Exceptions;
 using MediatR;
 using AuthConstants = Digdir.Domain.Dialogporten.Application.Common.Authorization.Constants;
@@ -11,12 +11,12 @@ namespace Digdir.Domain.Dialogporten.Application.Common.Behaviours;
 internal sealed class SilentUpdateBeaviour<TRequest, TResponse> : IPipelineBehavior<TRequest, TResponse>
     where TRequest : IRequest<TResponse>, ISilentUpdater
 {
-    private readonly ISilentUpdateContext _silentUpdateContext;
+    private readonly IApplicationContext _applicationContext;
     private readonly IUserResourceRegistry _userResourceRegistry;
 
-    public SilentUpdateBeaviour(ISilentUpdateContext silentUpdateContext, IUserResourceRegistry userResourceRegistry)
+    public SilentUpdateBeaviour(IApplicationContext applicationContext, IUserResourceRegistry userResourceRegistry)
     {
-        _silentUpdateContext = silentUpdateContext;
+        _applicationContext = applicationContext;
         _userResourceRegistry = userResourceRegistry;
     }
 
@@ -32,7 +32,7 @@ internal sealed class SilentUpdateBeaviour<TRequest, TResponse> : IPipelineBehav
 
         if (request.IsSilentUpdate)
         {
-            _silentUpdateContext.AddMetadata(Constants.IsSilentUpdate, bool.TrueString);
+            _applicationContext.AddMetadata(Constants.IsSilentUpdate, bool.TrueString);
         }
 
         return await next();
