+++ src/Digdir.Domain.Dialogporten.Application/Common/Context/ApplicationContext.cs
@@ -1,6 +1,6 @@
-namespace Digdir.Domain.Dialogporten.Domain.Common.DomainEvents;
+namespace Digdir.Domain.Dialogporten.Application.Common.Context;
 
-public sealed class SilentUpdateContext : ISilentUpdateContext
+public sealed class ApplicationContext : IApplicationContext
 {
     public Dictionary<string, string> Metadata { get; } = [];
     public void AddMetadata(string key, string value) => Metadata[key] = value;
